[Refactor] AppDelegate 내 Firestore 관련 import 제거, AppDelegate 관련없는 로직 각 파일로 분리
