+++ Project_Timer/Global/Extension/UIImage+Extension.swift
@@ -1,0 +1,19 @@
+//
+//  UIImage+Extension.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2022/07/03.
+//  Copyright © 2022 FDEE. All rights reserved.
+//
+
+import UIKit
+
+extension UIImage {
+    convenience init(view: UIView) {
+        UIGraphicsBeginImageContextWithOptions(view.bounds.size, view.isOpaque, 0.0)
+        view.drawHierarchy(in: view.bounds, afterScreenUpdates: false)
+        let image = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        self.init(cgImage: (image?.cgImage)!)
+    }
+}
