+++ Project_Timer/Global/Extension/UITextField+Extension.swift
@@ -1,0 +1,21 @@
+//
+//  UITextField+Extension.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2022/07/03.
+//  Copyright © 2022 FDEE. All rights reserved.
+//
+
+import UIKit
+
+extension UITextField {
+    func underlined() {
+        let border = CALayer()
+        let width = CGFloat(1.0)
+        border.borderColor = UIColor.lightGray.cgColor
+        border.frame = CGRect(x: 0, y: self.frame.size.height - width, width:  self.frame.size.width, height: self.frame.size.height)
+        border.borderWidth = width
+        self.layer.addSublayer(border)
+        self.layer.masksToBounds = true
+    }
+}
