+++ Project_Timer/Global/Extension/UIViewController+Extension.swift
@@ -9,6 +9,16 @@
 import UIKit
 
 extension UIViewController {
+    func hideKeyboard() {
+        let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self,
+            action: #selector(UIViewController.dismissKeyboard))
+        view.addGestureRecognizer(tap)
+    }
+    
+    @objc func dismissKeyboard() {
+        view.endEditing(true)
+    }
+    
     func showAlertWithOK(title: String, text: String) {
         let alert = UIAlertController(title: title, message: text, preferredStyle: .alert)
         let ok = UIAlertAction(title: "OK", style: .default)
