+++ Project_Timer/Global/Extension/UserDefaults+Extension.swift
@@ -1,0 +1,27 @@
+//
+//  UserDefaults+Extension.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2022/07/03.
+//  Copyright © 2022 FDEE. All rights reserved.
+//
+
+import UIKit
+
+extension UserDefaults {
+    func colorForKey(key: String) -> UIColor? {
+        var color: UIColor?
+        if let colorData = data(forKey: key) {
+            color = NSKeyedUnarchiver.unarchiveObject(with: colorData) as? UIColor
+        }
+        return color
+    }
+    
+    func setColor(color: UIColor?, forKey key: String) {
+        var colorData: NSData?
+        if let color = color {
+            colorData = NSKeyedArchiver.archivedData(withRootObject: color) as NSData?
+        }
+        set(colorData, forKey: key)
+    }
+}
